Improvements on the detection process and on the output video
Changed the process of generating the final line, calculating the average of the best lines and calculating the average line of a number of frames. Deleted the restrictions on the area of the frame to be considered when generating the lines. Started painting the area between the average lines in the final image and generated a new output video that contains all the important frames in one video.